Restructure some files
