+++ Resources/architecture-outline.md
